@@ -5,7 +5,7 @@
 // 2011-12-12 GONG Chen <chen.sst@gmail.com>
 //
 #include <boost/algorithm/string.hpp>
-#include <boost/filesystem.hpp>
+#include <filesystem>
 #include <stdint.h>
 #include <utf8.h>
 #include <utility>
@@ -27,7 +27,7 @@
 
 #ifdef _MSC_VER
 #include <opencc/UTF8Util.hpp>
-namespace fs = boost::filesystem;
+namespace fs = std::filesystem;
 #endif
 
 static const char* quote_left = "\xe3\x80\x94";   //"\xef\xbc\x88";
@@ -178,7 +178,7 @@ Simplifier::Simplifier(const Ticket& ticket)
 }
 
 void Simplifier::Initialize() {
-  using namespace boost::filesystem;
+  using namespace std::filesystem;
   initialized_ = true;  // no retry
   path opencc_config_path = opencc_config_;
   if (opencc_config_path.extension().string() == ".ini") {
