@@ -5,13 +5,12 @@
 // 2011-12-12 GONG Chen <chen.sst@gmail.com>
 //
 #include <stdint.h>
-#include <boost/lexical_cast.hpp>
 #include <rime/common.h>
 #include <rime/config.h>
 #include <rime/algo/utilities.h>
 #include <rime/lever/customizer.h>
 
-namespace fs = boost::filesystem;
+namespace fs = std::filesystem;
 
 namespace rime {
 
@@ -73,7 +72,7 @@ bool Customizer::UpdateConfigFile() {
   }
   string customization;
   if (!custom_path.empty() && fs::exists(custom_path)) {
-    customization = boost::lexical_cast<string>(Checksum(custom_path.string()));
+    customization = std::to_string(Checksum(custom_path.string()));
   }
   if (applied_customization != customization) {
     need_update = true;
