@@ -4,13 +4,13 @@
 //
 
 #include <boost/algorithm/string.hpp>
-#include <boost/filesystem.hpp>
+#include <filesystem>
 #include <rime/resource.h>
 
 namespace rime {
 
 string ResourceResolver::ToResourceId(const string& file_path) const {
-  string path_string = boost::filesystem::path(file_path).generic_string();
+  string path_string = std::filesystem::path(file_path).generic_string();
   bool has_prefix = boost::starts_with(path_string, type_.prefix);
   bool has_suffix = boost::ends_with(path_string, type_.suffix);
   size_t start = (has_prefix ? type_.prefix.length() : 0);
@@ -19,7 +19,7 @@ string ResourceResolver::ToResourceId(const string& file_path) const {
 }
 
 string ResourceResolver::ToFilePath(const string& resource_id) const {
-  boost::filesystem::path file_path(resource_id);
+  std::filesystem::path file_path(resource_id);
   bool missing_prefix = !file_path.has_parent_path() &&
                         !boost::starts_with(resource_id, type_.prefix);
   bool missing_suffix = !boost::ends_with(resource_id, type_.suffix);
@@ -27,21 +27,20 @@ string ResourceResolver::ToFilePath(const string& resource_id) const {
          (missing_suffix ? type_.suffix : "");
 }
 
-boost::filesystem::path ResourceResolver::ResolvePath(
-    const string& resource_id) {
-  return boost::filesystem::absolute(
-      boost::filesystem::path(type_.prefix + resource_id + type_.suffix),
-      root_path_);
+std::filesystem::path ResourceResolver::ResolvePath(const string& resource_id) {
+  return std::filesystem::absolute(
+      root_path_ /
+      std::filesystem::path(type_.prefix + resource_id + type_.suffix));
 }
 
-boost::filesystem::path FallbackResourceResolver::ResolvePath(
+std::filesystem::path FallbackResourceResolver::ResolvePath(
     const string& resource_id) {
   auto default_path = ResourceResolver::ResolvePath(resource_id);
-  if (!boost::filesystem::exists(default_path)) {
-    auto fallback_path = boost::filesystem::absolute(
-        boost::filesystem::path(type_.prefix + resource_id + type_.suffix),
-        fallback_root_path_);
-    if (boost::filesystem::exists(fallback_path)) {
+  if (!std::filesystem::exists(default_path)) {
+    auto fallback_path = std::filesystem::absolute(
+        fallback_root_path_ /
+        std::filesystem::path(type_.prefix + resource_id + type_.suffix));
+    if (std::filesystem::exists(fallback_path)) {
       return fallback_path;
     }
   }
