@@ -6,7 +6,7 @@
 #ifndef RIME_RESOURCE_H_
 #define RIME_RESOURCE_H_
 
-#include <boost/filesystem.hpp>
+#include <filesystem>
 #include <rime_api.h>
 #include <rime/common.h>
 
@@ -24,17 +24,17 @@ class RIME_API ResourceResolver {
  public:
   explicit ResourceResolver(const ResourceType type) : type_(type) {}
   virtual ~ResourceResolver() {}
-  virtual boost::filesystem::path ResolvePath(const string& resource_id);
+  virtual std::filesystem::path ResolvePath(const string& resource_id);
   string ToResourceId(const string& file_path) const;
   string ToFilePath(const string& resource_id) const;
-  void set_root_path(boost::filesystem::path root_path) {
+  void set_root_path(std::filesystem::path root_path) {
     root_path_ = root_path;
   }
-  boost::filesystem::path root_path() const { return root_path_; }
+  std::filesystem::path root_path() const { return root_path_; }
 
  protected:
   const ResourceType type_;
-  boost::filesystem::path root_path_;
+  std::filesystem::path root_path_;
 };
 
 // try fallback path if target file doesn't exist in root path
@@ -42,13 +42,13 @@ class RIME_API FallbackResourceResolver : public ResourceResolver {
  public:
   explicit FallbackResourceResolver(const ResourceType& type)
       : ResourceResolver(type) {}
-  boost::filesystem::path ResolvePath(const string& resource_id) override;
-  void set_fallback_root_path(boost::filesystem::path fallback_root_path) {
+  std::filesystem::path ResolvePath(const string& resource_id) override;
+  void set_fallback_root_path(std::filesystem::path fallback_root_path) {
     fallback_root_path_ = fallback_root_path;
   }
 
  private:
-  boost::filesystem::path fallback_root_path_;
+  std::filesystem::path fallback_root_path_;
 };
 
 }  // namespace rime
