@@ -5,14 +5,15 @@
 // 2012-07-07 GONG Chen <chen.sst@gmail.com>
 //
 #include <iostream>
-#include <boost/filesystem.hpp>
+#include <filesystem>
 #include <rime/config.h>
 #include <rime/deployer.h>
 #include <rime/service.h>
 #include <rime/setup.h>
 #include <rime/lever/deployment_tasks.h>
+#include "codepage.h"
 
-namespace fs = boost::filesystem;
+namespace fs = std::filesystem;
 
 using namespace rime;
 
@@ -36,7 +37,8 @@ int add_schema(int count, char* schemas[]) {
         break;
       }
     }
-    if (already_there) continue;
+    if (already_there)
+      continue;
     schema_list[schema_list.size()]["schema"] = new_schema_id;
     LOG(INFO) << "added schema: " << new_schema_id;
   }
@@ -60,8 +62,7 @@ int set_active_schema(const string& schema_id) {
   return 0;
 }
 
-static void setup_deployer(Deployer* deployer,
-                           int argc, char* argv[]) {
+static void setup_deployer(Deployer* deployer, int argc, char* argv[]) {
   if (argc > 0) {
     deployer->user_data_dir = argv[0];
   }
@@ -73,39 +74,52 @@ static void setup_deployer(Deployer* deployer,
   if (argc > 2) {
     deployer->staging_dir = argv[2];
   } else {
-    deployer->staging_dir = (fs::path(deployer->user_data_dir) / "build").string();
+    deployer->staging_dir =
+        (fs::path(deployer->user_data_dir) / "build").string();
   }
-  deployer->prebuilt_data_dir = (fs::path(deployer->shared_data_dir) / "build").string();
+  deployer->prebuilt_data_dir =
+      (fs::path(deployer->shared_data_dir) / "build").string();
 }
 
 int main(int argc, char* argv[]) {
+  unsigned int codepage = SetConsoleOutputCodePage();
   SetupLogging("rime.tools");
 
   if (argc == 1) {
-    std::cout << "Usage: " << std::endl
-              << "\t--add-schema <schema_id>..." << std::endl
-              << "\t\tAdd one or more schema_id(s) to the schema_list, write patch in default.custom.yaml" << std::endl
-              << std::endl
-              << "\t--build [user_data_dir] [shared_data_dir] [staging_dir]" << std::endl
-              << "\t\tBuild and deploy Rime data." << std::endl
-              << "\t\tIf unspecified, user_data_dir and shared_data_dir defaults to the working directory." << std::endl
-              << "\t\tTo deploy data for ibus-rime, use the following directories:" << std::endl
-              << "\t\tuser_data_dir    ~/.config/ibus/rime" << std::endl
-              << "\t\tshared_data_dir  /usr/share/rime-data" << std::endl
-              << "\t\tstaging_dir      ~/.config/ibus/rime/build" << std::endl
-              << std::endl
-              << "\t--compile <x.schema.yaml> [user_data_dir] [shared_data_dir] [staging_dir]" << std::endl
-              << "\t\tCompile a specific schema's dictionary files." << std::endl
-              << std::endl
-              << "\t--set-active-schema <schema_id>" << std::endl
-              << "\t\tSet the active schema in user.yaml" << std::endl
-      ;
+    std::cout
+        << "Usage: " << std::endl
+        << "\t--add-schema <schema_id>..." << std::endl
+        << "\t\tAdd one or more schema_id(s) to the schema_list, write patch "
+           "in default.custom.yaml"
+        << std::endl
+        << std::endl
+        << "\t--build [user_data_dir] [shared_data_dir] [staging_dir]"
+        << std::endl
+        << "\t\tBuild and deploy Rime data." << std::endl
+        << "\t\tIf unspecified, user_data_dir and shared_data_dir defaults to "
+           "the working directory."
+        << std::endl
+        << "\t\tTo deploy data for ibus-rime, use the following directories:"
+        << std::endl
+        << "\t\tuser_data_dir    ~/.config/ibus/rime" << std::endl
+        << "\t\tshared_data_dir  /usr/share/rime-data" << std::endl
+        << "\t\tstaging_dir      ~/.config/ibus/rime/build" << std::endl
+        << std::endl
+        << "\t--compile <x.schema.yaml> [user_data_dir] [shared_data_dir] "
+           "[staging_dir]"
+        << std::endl
+        << "\t\tCompile a specific schema's dictionary files." << std::endl
+        << std::endl
+        << "\t--set-active-schema <schema_id>" << std::endl
+        << "\t\tSet the active schema in user.yaml" << std::endl;
 
+    SetConsoleOutputCodePage(codepage);
     return 0;
   }
 
   string option;
-  if (argc >= 2) option = argv[1];
+  if (argc >= 2)
+    option = argv[1];
   // shift
   argc -= 2, argv += 2;
 
@@ -114,15 +128,21 @@ int main(int argc, char* argv[]) {
     setup_deployer(&deployer, argc, argv);
     LoadModules(kDeployerModules);
     WorkspaceUpdate update;
-    return update.Run(&deployer) ? 0 : 1;
+    int res = update.Run(&deployer) ? 0 : 1;
+    SetConsoleOutputCodePage(codepage);
+    return res;
   }
 
   if (argc >= 1 && option == "--add-schema") {
-    return add_schema(argc, argv);
+    int res = add_schema(argc, argv);
+    SetConsoleOutputCodePage(codepage);
+    return res;
   }
 
   if (argc == 1 && option == "--set-active-schema") {
-    return set_active_schema(argv[0]);
+    int res = set_active_schema(argv[0]);
+    SetConsoleOutputCodePage(codepage);
+    return res;
   }
 
   if (argc >= 1 && option == "--compile") {
@@ -132,9 +152,12 @@ int main(int argc, char* argv[]) {
     string schema_file(argv[0]);
     SchemaUpdate update(schema_file);
     update.set_verbose(true);
-    return update.Run(&deployer) ? 0 : 1;
+    int res = update.Run(&deployer) ? 0 : 1;
+    SetConsoleOutputCodePage(codepage);
+    return res;
   }
 
   std::cerr << "invalid arguments." << std::endl;
+  SetConsoleOutputCodePage(codepage);
   return 1;
 }
