@@ -5,7 +5,7 @@
 // 2014-12-04 Chen Gong <chen.sst@gmail.com>
 //
 
-#include <boost/filesystem.hpp>
+#include <filesystem>
 #include <leveldb/db.h>
 #include <leveldb/write_batch.h>
 #include <rime/common.h>
