@@ -4,8 +4,7 @@
 //
 // 2011-11-27 GONG Chen <chen.sst@gmail.com>
 //
-#include <boost/filesystem.hpp>
-#include <boost/lexical_cast.hpp>
+#include <filesystem>
 #include <utf8.h>
 #include <rime/resource.h>
 #include <rime/service.h>
@@ -74,7 +73,7 @@ bool PresetVocabulary::GetWeightForEntry(const string& key, double* weight) {
   if (!db_ || !db_->Fetch(key, &weight_str))
     return false;
   try {
-    *weight = boost::lexical_cast<double>(weight_str);
+    *weight = std::stod(weight_str);
   } catch (...) {
     return false;
   }
@@ -105,7 +104,7 @@ bool PresetVocabulary::IsQualifiedPhrase(const string& phrase,
       return false;
   }
   if (min_phrase_weight_ > 0.0) {
-    double weight = boost::lexical_cast<double>(weight_str);
+    double weight = std::stod(weight_str);
     if (weight < min_phrase_weight_)
       return false;
   }
