@@ -9,8 +9,7 @@
 #include <cstdlib>
 #include <sstream>
 #include <boost/algorithm/string.hpp>
-#include <boost/filesystem.hpp>
-#include <boost/lexical_cast.hpp>
+#include <filesystem>
 #include <rime/resource.h>
 #include <rime/schema.h>
 #include <rime/service.h>
