@@ -5,7 +5,6 @@
 // 2013-04-18 GONG Chen <chen.sst@gmail.com>
 //
 #include <boost/algorithm/string.hpp>
-#include <boost/lexical_cast.hpp>
 #include <rime/dict/table_db.h>
 #include <rime/dict/user_db.h>
 
@@ -28,7 +27,7 @@ static bool rime_table_entry_parser(const Tsv& row,
   UserDbValue v;
   if (row.size() >= 3 && !row[2].empty()) {
     try {
-      v.commits = boost::lexical_cast<int>(row[2]);
+      v.commits = std::stoi(row[2]);
       const double kS = 1e8;
       v.dee = (v.commits + 1) / kS;
     } catch (...) {
@@ -51,7 +50,7 @@ static bool rime_table_entry_formatter(const string& key,
     return false;
   boost::algorithm::trim(row[0]);  // remove trailing space
   row[0].swap(row[1]);
-  row.push_back(boost::lexical_cast<string>(v.commits));
+  row.push_back(std::to_string(v.commits));
   return true;
 }
 
