@@ -8,22 +8,28 @@
 #include <stdlib.h>
 #include <string.h>
 #include <rime_api.h>
+#include "codepage.h"
 
-void print_status(RimeStatus *status) {
-  printf("schema: %s / %s\n",
-         status->schema_id, status->schema_name);
+void print_status(RimeStatus* status) {
+  printf("schema: %s / %s\n", status->schema_id, status->schema_name);
   printf("status: ");
-  if (status->is_disabled) printf("disabled ");
-  if (status->is_composing) printf("composing ");
-  if (status->is_ascii_mode) printf("ascii ");
-  if (status->is_full_shape) printf("full_shape ");
-  if (status->is_simplified) printf("simplified ");
+  if (status->is_disabled)
+    printf("disabled ");
+  if (status->is_composing)
+    printf("composing ");
+  if (status->is_ascii_mode)
+    printf("ascii ");
+  if (status->is_full_shape)
+    printf("full_shape ");
+  if (status->is_simplified)
+    printf("simplified ");
   printf("\n");
 }
 
-void print_composition(RimeComposition *composition) {
-  const char *preedit = composition->preedit;
-  if (!preedit) return;
+void print_composition(RimeComposition* composition) {
+  const char* preedit = composition->preedit;
+  if (!preedit)
+    return;
   size_t len = strlen(preedit);
   size_t start = composition->sel_start;
   size_t end = composition->sel_end;
@@ -36,37 +42,34 @@ void print_composition(RimeComposition *composition) {
         putchar(']');
       }
     }
-    if (i == cursor) putchar('|');
+    if (i == cursor)
+      putchar('|');
     if (i < len)
-        putchar(preedit[i]);
+      putchar(preedit[i]);
   }
   printf("\n");
 }
 
-void print_menu(RimeMenu *menu) {
-  if (menu->num_candidates == 0) return;
-  printf("page: %d%c (of size %d)\n",
-         menu->page_no + 1,
-         menu->is_last_page ? '$' : ' ',
-         menu->page_size);
+void print_menu(RimeMenu* menu) {
+  if (menu->num_candidates == 0)
+    return;
+  printf("page: %d%c (of size %d)\n", menu->page_no + 1,
+         menu->is_last_page ? '$' : ' ', menu->page_size);
   for (int i = 0; i < menu->num_candidates; ++i) {
     bool highlighted = i == menu->highlighted_candidate_index;
-    printf("%d. %c%s%c%s\n",
-           i + 1,
-           highlighted ? '[' : ' ',
-           menu->candidates[i].text,
-           highlighted ? ']' : ' ',
+    printf("%d. %c%s%c%s\n", i + 1, highlighted ? '[' : ' ',
+           menu->candidates[i].text, highlighted ? ']' : ' ',
            menu->candidates[i].comment ? menu->candidates[i].comment : "");
   }
 }
 
-void print_context(RimeContext *context) {
+void print_context(RimeContext* context) {
   if (context->composition.length > 0) {
     print_composition(&context->composition);
-    print_menu(&context->menu);
   } else {
     printf("(not composing)\n");
   }
+  print_menu(&context->menu);
 }
 
 void print(RimeSessionId session_id) {
@@ -99,8 +102,8 @@ bool execute_special_command(const char* line, RimeSessionId session_id) {
     if (rime->get_schema_list(&list)) {
       printf("schema list:\n");
       for (size_t i = 0; i < list.size; ++i) {
-        printf("%lu. %s [%s]\n", (i + 1),
-               list.list[i].name, list.list[i].schema_id);
+        printf("%lu. %s [%s]\n", (i + 1), list.list[i].name,
+               list.list[i].schema_id);
       }
       rime->free_schema_list(&list);
     }
@@ -175,12 +178,14 @@ void on_message(void* context_object,
     const char* state_label =
         rime->get_state_label(session_id, option_name, state);
     if (state_label) {
-      printf("updated option: %s = %d // %s\n", option_name, state, state_label);
+      printf("updated option: %s = %d // %s\n", option_name, state,
+             state_label);
     }
   }
 }
 
-int main(int argc, char *argv[]) {
+int main(int argc, char* argv[]) {
+  unsigned int codepage = SetConsoleOutputCodePage();
   RimeApi* rime = rime_get_api();
 
   RIME_STRUCT(RimeTraits, traits);
@@ -190,6 +195,7 @@ int main(int argc, char *argv[]) {
   rime->set_notification_handler(&on_message, NULL);
 
   fprintf(stderr, "initializing...\n");
+reload:
   rime->initialize(NULL);
   Bool full_check = True;
   if (rime->start_maintenance(full_check))
@@ -199,13 +205,14 @@ int main(int argc, char *argv[]) {
   RimeSessionId session_id = rime->create_session();
   if (!session_id) {
     fprintf(stderr, "Error creating rime session.\n");
+    SetConsoleOutputCodePage(codepage);
     return 1;
   }
 
   const int kMaxLength = 99;
   char line[kMaxLength + 1] = {0};
   while (fgets(line, kMaxLength, stdin) != NULL) {
-    for (char *p = line; *p; ++p) {
+    for (char* p = line; *p; ++p) {
       if (*p == '\r' || *p == '\n') {
         *p = '\0';
         break;
@@ -213,6 +220,11 @@ int main(int argc, char *argv[]) {
     }
     if (!strcmp(line, "exit"))
       break;
+    else if (!strcmp(line, "reload")) {
+      rime->destroy_session(session_id);
+      rime->finalize();
+      goto reload;
+    }
     if (execute_special_command(line, session_id))
       continue;
     if (rime->simulate_key_sequence(session_id, line)) {
@@ -226,5 +238,6 @@ int main(int argc, char *argv[]) {
 
   rime->finalize();
 
+  SetConsoleOutputCodePage(codepage);
   return 0;
 }
